+++ test/tests.js
@@ -4,23 +4,33 @@ describe('history', function () {
 var assert = require('assert')
   , history = require('history');
 
-describe('get', function () {
-  it('should return the pathname', function () {
-    assert(location.pathname === history.get());
+describe('path', function () {
+  it('should return the current path', function () {
+    assert(window.location.pathname === history.path());
+  });
+});
+
+describe('state', function () {
+  it('should return the current state', function () {
+    var obj = { check: true };
+    window.history.pushState(obj, null, null);
+    assert(true === history.state().check);
   });
 });
 
 describe('push', function () {
   it('should push an entry onto the history', function () {
-    history.push('push');
-    assert('/push' === history.get());
+    history.push('push', { method: 'push' });
+    assert('/push' === history.path());
+    assert('push' === history.state().method);
   });
 });
 
 describe('replace', function () {
   it('should replace the current entry in the history', function () {
-    history.replace('replace');
-    assert('/replace' === history.get());
+    history.replace('replace', { method: 'replace' });
+    assert('/replace' === history.path());
+    assert('replace' === history.state().method);
   });
 });
 
@@ -28,10 +38,10 @@ describe('back', function () {
   it('should move backwards', function (done) {
     history.push('1');
     history.push('2');
-    assert('/2' === history.get());
+    assert('/2' === history.path());
     history.back();
     setTimeout(function () {
-      assert('/1' === history.get());
+      assert('/1' === history.path());
       done();
     }, 200);
   });
@@ -40,10 +50,10 @@ describe('back', function () {
     history.push('1');
     history.push('2');
     history.push('3');
-    assert('/3' === history.get());
+    assert('/3' === history.path());
     history.back(2);
     setTimeout(function () {
-      assert('/1' === history.get());
+      assert('/1' === history.path());
       done();
     }, 200);
   });
@@ -56,7 +66,7 @@ describe('forward', function () {
     history.back();
     history.forward();
     setTimeout(function () {
-      assert('/2' === history.get());
+      assert('/2' === history.path());
       done();
     }, 200);
   });
@@ -68,7 +78,7 @@ describe('forward', function () {
     history.back(2);
     history.forward(2);
     setTimeout(function () {
-      assert('/3' === history.get());
+      assert('/3' === history.path());
       done();
     }, 200);
   });