+++ History.md
@@ -1,3 +1,9 @@
+0.0.2 - July 25, 2013
+---------------------
+* added `state` argument to `#push` and `#replace`
+* added `#state` getter
+* changed path getter to `#path`
+
 0.0.1 - July 9, 2013
 --------------------
 :sparkles: